Adds support for lists returned by selectAssoc
The column values parameter to RippingCluster_Database::selectAssoc()
now accepts an array of columns instead of a single value. If a scalar
is provided, the value is returned as a simple associative array; if an
array is provided, the full list of columns requested is returned as a
nested associative array.