Adds datatype support to the Config class
The config class now supports a type column in the database which
specifies how the value should be interpreted. The class defines a
number of scalar types, which are returned without transformation. A
string list type is also defined, which splits the value into an array
based on newlines.

Also added phpdoc comments to all class constants, members and methods.